requirements.txt
boto3==1.24.70
python-dotenv==0.21.0
pyTelegramBotAPI==4.4.1
uuid==1.30
